+++ src/tc/azext_tc/_completers.py
@@ -6,7 +6,7 @@
 from azure.cli.core.decorators import Completer
 
 from knack.log import get_logger
-# from knack.util import CLIError
+
 from ._client_factory import teamcloud_client_factory
 
 logger = get_logger(__name__)
@@ -23,5 +23,35 @@ def get_project_completion_list(cmd, prefix, namespace, **kwargs):  # pylint: di
 
     try:
         return [p.name for p in result.data]
+    except AttributeError:
+        return []
+
+
+@Completer
+def get_project_type_completion_list(cmd, prefix, namespace, **kwargs):  # pylint: disable=unused-argument
+
+    # logger.warning('Completing...')
+
+    client = teamcloud_client_factory(cmd.cli_ctx)
+    client._client.config.base_url = namespace.base_url  # pylint: disable=protected-access
+    result = client.get_project_types()
+
+    try:
+        return [p.id for p in result.data]
+    except AttributeError:
+        return []
+
+
+@Completer
+def get_provider_completion_list(cmd, prefix, namespace, **kwargs):  # pylint: disable=unused-argument
+
+    # logger.warning('Completing...')
+
+    client = teamcloud_client_factory(cmd.cli_ctx)
+    client._client.config.base_url = namespace.base_url  # pylint: disable=protected-access
+    result = client.get_providers()
+
+    try:
+        return [p.id for p in result.data]
     except AttributeError:
         return []