+++ src/tc/azext_tc/commands.py
@@ -26,8 +26,10 @@ def load_command_table(self, _):
     # TeamCloud Users
 
     with self.command_group('tc user', client_factory=teamcloud_client_factory) as g:
-        g.custom_command('create', 'teamcloud_user_create', transform=transform_output)  # no_wait
-        g.custom_command('delete', 'teamcloud_user_delete', transform=transform_output)  # no_wait
+        g.custom_command('create', 'teamcloud_user_create', transform=transform_output,
+                         supports_no_wait=True)
+        g.custom_command('delete', 'teamcloud_user_delete', transform=transform_output,
+                         supports_no_wait=True, confirmation='Are you sure you want to delete this user?')
         g.custom_command('list', 'teamcloud_user_list', transform=transform_output,
                          table_transformer=transform_user_table_output)
         g.custom_show_command('show', 'teamcloud_user_get', transform=transform_output)
@@ -35,8 +37,10 @@ def load_command_table(self, _):
     # TeamCloud Tags
 
     with self.command_group('tc tag', client_factory=teamcloud_client_factory) as g:
-        g.custom_command('create', 'teamcloud_tag_create', transform=transform_output)  # no_wait
-        g.custom_command('delete', 'teamcloud_tag_delete', transform=transform_output)  # no_wait
+        g.custom_command('create', 'teamcloud_tag_create', transform=transform_output,
+                         supports_no_wait=True)
+        g.custom_command('delete', 'teamcloud_tag_delete', transform=transform_output,
+                         supports_no_wait=True)
         g.custom_command('list', 'teamcloud_tag_list', transform=transform_output,
                          table_transformer=transform_tag_table_output)
         g.custom_show_command('show', 'teamcloud_tag_get', transform=transform_output)
@@ -44,7 +48,8 @@ def load_command_table(self, _):
     # Projects
 
     with self.command_group('tc project', client_factory=teamcloud_client_factory) as g:
-        g.custom_command('create', 'project_create', transform=transform_output)  # no_wait
+        g.custom_command('create', 'project_create', transform=transform_output,
+                         supports_no_wait=True)
         g.custom_command('delete', 'project_delete', transform=transform_output,
                          supports_no_wait=True, confirmation='Are you sure you want to delete this project?')
         g.custom_command('list', 'project_list', transform=transform_output,
@@ -54,8 +59,10 @@ def load_command_table(self, _):
     # Project Users
 
     with self.command_group('tc project user', client_factory=teamcloud_client_factory) as g:
-        g.custom_command('create', 'project_user_create', transform=transform_output)  # no_wait
-        g.custom_command('delete', 'project_user_delete', transform=transform_output)  # no_wait
+        g.custom_command('create', 'project_user_create', transform=transform_output,
+                         supports_no_wait=True)
+        g.custom_command('delete', 'project_user_delete', transform=transform_output,
+                         supports_no_wait=True, confirmation='Are you sure you want to delete this user?')
         g.custom_command('list', 'project_user_list', transform=transform_output,
                          table_transformer=transform_user_table_output)
         g.custom_show_command('show', 'project_user_get', transform=transform_output)
@@ -63,8 +70,10 @@ def load_command_table(self, _):
     # Project Tags
 
     with self.command_group('tc project tag', client_factory=teamcloud_client_factory) as g:
-        g.custom_command('create', 'project_tag_create', transform=transform_output)  # no_wait
-        g.custom_command('delete', 'project_tag_delete', transform=transform_output)  # no_wait
+        g.custom_command('create', 'project_tag_create', transform=transform_output,
+                         supports_no_wait=True)
+        g.custom_command('delete', 'project_tag_delete', transform=transform_output,
+                         supports_no_wait=True)
         g.custom_command('list', 'project_tag_list', transform=transform_output,
                          table_transformer=transform_tag_table_output)
         g.custom_show_command('show', 'project_tag_get', transform=transform_output)
@@ -73,7 +82,8 @@ def load_command_table(self, _):
 
     with self.command_group('tc project-type', client_factory=teamcloud_client_factory) as g:
         g.custom_command('create', 'project_type_create', transform=transform_output)
-        g.custom_command('delete', 'project_type_delete', transform=transform_output)  # confirm?
+        g.custom_command('delete', 'project_type_delete', transform=transform_output,
+                         confirmation='Are you sure you want to delete this project type?')
         g.custom_command('list', 'project_type_list', transform=transform_output,
                          table_transformer=transform_project_type_table_output)
         g.custom_show_command('show', 'project_type_get', transform=transform_output)
@@ -81,8 +91,10 @@ def load_command_table(self, _):
     # Providers
 
     with self.command_group('tc provider', client_factory=teamcloud_client_factory) as g:
-        g.custom_command('create', 'provider_create', transform=transform_output)
-        g.custom_command('delete', 'provider_delete', transform=transform_output)  # confirm?
+        g.custom_command('create', 'provider_create', transform=transform_output,
+                         supports_no_wait=True)
+        g.custom_command('delete', 'provider_delete', transform=transform_output,
+                         supports_no_wait=True, confirmation='Are you sure you want to delete this provider?')
         g.custom_command('list', 'provider_list', transform=transform_output,
                          table_transformer=transform_provider_table_output)
         g.custom_show_command('show', 'provider_get', transform=transform_output)